pyproject.toml
[project]
name = "dreamcatcher-backend"
version = "0.1.0"
description = "DreamCatcher backend API: serves the DTSC.ipynb pipeline (embed -> archetype -> emotions) as a REST endpoint."
requires-python = ">=3.10,<3.13"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.7",
    "sentence-transformers>=2.7",
    "umap-learn>=0.5.5",
    "hdbscan>=0.8.33",
    "scikit-learn>=1.3",
    "transformers>=4.41",
    "torch>=2.2",
    "nltk>=3.8",
    "pandas>=2.1",
    "numpy>=1.26,<2.0",
    "joblib>=1.3",
    "pyarrow>=15.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["dreamcatcher*"]
